Manually generate distribution archives (#1032)
Github has a habit of changing how the automatic source archives are generated (e.g. compression), which
changes their checksums, which then breaks users. Instead of risking that happening again, generate
the distribution archive as part of the release workflow and upload it to the release. The same mechanism,
`git export` is used, so the format and structure is the same as before.