docs: Simplify pull request template (#1100)
The existing template is very verbose with many of the lines not being
applicable to any given PR. This also makes the PR description poorly
suitable for a commit description.

By having a cleaner PR description that is commit message friendly,
whoever merges the PR can more easily create a meaningful commit
message. It also allows other maintainer's to clean up the description
prior to it being merged.