feat: bzl file per rule/provider (#1122)
This is basically a performance optimization. Bazel can only see that a
bzl file changed, not the particular contents. This means that any
downstream bzl file loading it is invalidated, even if it doesn't load
any of the affected code. As an example, if a package only loads
`py_library.bzl`, then changing `py_test.bzl` doesn't need to invalidate
all libraries.

* Also removes some more extraneous license comments

Work towards #1069